# api/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: marc.sturm@med.uni-tuebingen.de
  description: Short-read sequencing tools
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: ngs-remote
  version: "2018_11"
externalDocs:
  description: Find out more about ngs-bits
  url: http://github.com/imgag/ngs-bits
servers:
- url: /v1
security:
- jwt:
  - secret
tags:
- description: Various tools depending on ngs-bits
  name: ngs-bits
- description: Various tools depending on megSAP
  name: megSAP
- description: Controller dealing with file related issues
  name: files
paths:
  /VariantFilterAnnotations:
    post:
      operationId: variant_filter_annotations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariantFilterRequest'
      responses:
        200:
          description: successfull
        400:
          content:
            application/json:
              schema:
                type: string
          description: bad request
      tags:
      - ngs-bits
      x-openapi-router-controller: openapi_server.controllers.ngs_bits_controller
  /VcfCheck/{filePath}:
    get:
      description: Check a file at given path with VcfCheck.
      operationId: vcf_check_file_path_get
      parameters:
      - description: Path to file
        explode: false
        in: path
        name: filePath
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          description: successfull
        400:
          content:
            application/json:
              schema:
                type: string
          description: error
      tags:
      - ngs-bits
      x-openapi-router-controller: openapi_server.controllers.ngs_bits_controller
  /an_vep/{filePath}:
    get:
      description: Annotates a VCF using VEP
      operationId: an_vep_file_path_get
      parameters:
      - description: The VCF file to annotate
        explode: false
        in: path
        name: filePath
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                type: number
          description: successfull
        400:
          content:
            application/json:
              schema:
                type: string
          description: error
      tags:
      - megSAP
      x-openapi-router-controller: openapi_server.controllers.meg_sap_controller
  /count/{filePath}:
    get:
      description: Count items in a file
      operationId: count_file_path_get
      parameters:
      - description: Path to the file
        explode: false
        in: path
        name: filePath
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                type: number
          description: successfull
      tags:
      - files
      x-openapi-router-controller: openapi_server.controllers.files_controller
  /download/{filePath}:
    get:
      description: Download a file from given path eventually cutting it with Lines
        header
      operationId: download_file_path_get
      parameters:
      - description: Path to the file
        explode: false
        in: path
        name: filePath
        required: true
        schema:
          type: string
        style: simple
      - description: The lines to serve represented as from-to e.g 1-100 starting
          at 1
        explode: false
        in: header
        name: Lines
        required: false
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            text/gsvar:
              schema:
                format: binary
                type: string
          description: successfull
      tags:
      - files
      x-openapi-router-controller: openapi_server.controllers.files_controller
  /upload:
    post:
      description: Uploads a file to the server
      operationId: upload_post
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          multipart/form-data:
            schema:
              properties:
                uploadedFile:
                  description: The GSVar file to upload.
                  format: binary
                  type: string
              type: object
      responses:
        200:
          description: successfull
        400:
          content:
            '*/*':
              schema:
                type: string
          description: failed, e.g file can not be uploaded
      tags:
      - files
      x-openapi-router-controller: openapi_server.controllers.files_controller
  /vcf2gsvar/{filePath}:
    get:
      description: Converts a VCF file to a GSvar
      operationId: vcf2gsvar_file_path_get
      parameters:
      - description: The GSvar file to convert
        explode: false
        in: path
        name: filePath
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                type: number
          description: successfull
        400:
          content:
            application/json:
              schema:
                type: string
          description: error
      tags:
      - megSAP
      x-openapi-router-controller: openapi_server.controllers.meg_sap_controller
components:
  requestBodies:
    inline_object:
      content:
        multipart/form-data:
          schema:
            $ref: '#/components/schemas/inline_object'
  schemas:
    VariantFilterRequest:
      example:
        filter:
        - '{}'
        - '{}'
        in: in
        out: out
      properties:
        in:
          format: file
          type: string
        out:
          format: file
          type: string
        filter:
          description: The filter dict to apply
          items:
            type: object
          type: array
      type: object
    inline_object:
      properties:
        uploadedFile:
          description: The GSVar file to upload.
          format: binary
          type: string
      type: object
  securitySchemes:
    jwt:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller_.info_from_jwt